backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
supabase==1.0.3
python-dotenv==0.19.0
numpy>=1.24.0
pandas==1.3.3
matplotlib==3.4.3
docker==6.1.3
pyjwt==2.3.0 